src/streamlit/requirements.txt
# Streamlit application dependencies
streamlit>=1.29.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0

# Strands Agents SDK - AWS open-source AI agent framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS Bedrock AgentCore SDK - Memory integration with Strands
bedrock-agentcore>=0.1.0
